checkbox improvement + correction for when $fields is empty
A poor assumption was made that if $fields was empty then the widget
shouldn't be displayed. First of all, doing so caused a PHP error as the
widget was still being registered with WordPress. Secondly, it is
entirely conceivable (as was in my case) that a widget be generated
without any options at all, yet still use this class.  The other change
has to do with the checkbox field output, showing the label below the
checkbox did not look very good, especially with the ":", and thus I
moved it above the field.